refactor(swift): split ComposeForm and EmailDetailView
Two files were over 800 lines after the first round of splits and
were flagged in the follow-up code-quality review:

- EmailDetailView.swift (1029 → 386): the ThreadMessageCardView
  struct already lived inside this file but was 642 lines on its
  own — moved to its own file. No access-modifier changes needed
  since it's a separate struct with no private cross-references.

- ComposeForm.swift (980 → 764):
  * AttachmentChip and VimSearchPill subviews (~80 lines) moved
    to ComposeFormSubviews.swift — both are self-contained with
    no ComposeForm state access.
  * Contact suggestion handling (handlePartialTextChange,
    handleSuggestionKeyCommand, selectContact,
    dismissContactSuggestions, currentTokensForField, ~137 lines)
    moved to ComposeForm+Contacts.swift as an extension. The six
    @State properties for contact popup state and scheduleAutoSave
    were changed from private to internal so the extension can
    reach them.

No behavioral changes.
